pyproject.toml
[tool.poetry]
name = "blombo"
version = "0.1.0"
description = "A context management system for LLMs"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "blombo", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.0.0"
google-auth-oauthlib = "^1.0.0"
google-auth = "^2.0.0"
google-api-python-client = "^2.0.0"
types-setuptools = "^57.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
ruff = "^0.2.1"
mypy = "^1.0.0"
isort = "^5.0.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "B", "I", "N", "UP", "ANN", "S", "A", "C4", "DTZ", "T10", "EM", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "RUF"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3 